build: update package.json to hoist deps and use file paths
The packages managed by monorepo used to refer to one another by version
instead of file reference in package.json. This required setup-dev to
run a lerna bootstrap command.

Lerna recommends to use file paths in package.json for packages managed
by the monorepo. This change coverts subpackages package.json to use the
recommendation and hoists all the dev dependencies to root of the monorepo
and seepds up the setup time.

This change also migrates to yarn as npm client as it supports for workspaces
- updated the npm client to yarn
- updated the packages to use local file path in mono repo packages
- updated lerna to use workspaces
- updated the setup-dev scripts