CI: Use setup-dev script for build (#399)
The CI script used to use product-build script to bootstrap the package. This script hoisted the dependencies to the monorepo root. In the integration tests, we run setup-dev script to make sure amplify command is available globally. But when the setup-dev script is run, it causes a conflict in what packages to use for building typescript packages.