fix(1342): api push error on CI/CD platforms. (#1383)
Fixes a bug related to pushing on CI/CD platforms. This change causes every push invocation to
update the S3DeploymentRootKey using a fixed hash function. This is in response to #1342. After
updating to this new version of the CLI, you will need to run `amplify push` one time with changes
to the api category and then every subsequent CI/CD push should succeed regardless of if you are
updating the api category or not.

re #1342