fix(amplify-category-auth): add policy name char length limit
In pr #1199 the name of coginto userpool was changed to include the project name and a random
number. We use the this to name to the resource and - char is not supported in resource name.
Updated the resource name not to include -. Updated the policyname to use substring of resource name
as there is a limit of 65 chars in policy name