fix(amplify-provider-awscloudformation): generate consistent s3 keys
Amplify CLI used to use date to ensure that the functions S3 key was changing every time it got
pushed to the cloud. This caused multi env with multi branch workflow to have git merge conflict as
these keys kept changing based on the time when the function gets pushed. Updated the code to use the
folder hash of function so the key is consistent if the content has not changed

fix #1666