feat(graphql-relation-transformer): added get item and style changes
wrote function to configure the resolver needed to run a getItem for when the relation is using the
default table and providing the full primary key. Added comments explaining directive purpose and
use, refactored the validate field function and updated the package.json (all based on pull request
comments from Paris). Added comments to functions that were left uncommented and rearranged some of
the code for stylistic reasons.