feat(graphql-relation-transformer): added query functionality
added code to configure resolvers to run dynamodb queries that return a list of objects. This
enables users to use relation to create relations to many objects rather than just one as with
getItem. Also added a test to check that the resolver has the right components and that that the
field is updated correctly.