fix(amplify-util-mock): prevent resolver file overwrite in windows (#2007)
Amplify mock creates the files that are generated by the GraphQL transformer to make it easy to customize them. If a resolver was customized and mock server was started on windows, the customized resolver file got overwritten with the generated file from the transformer on windows. This was caused due difference in the file path separator produced by transformer and filesystem. Fixed it by normalizing the file path

fix #2006