fix(amplify-util-mock): include custom resolver templates
Mock ignored custom resolver mapping template existing in the resolver folder. Transformer does not
include custom resolvers if they are not overriding the ones that are generated by the transformer.
This is the case when user has a custom stack with additional queries/mutations and resolvers.
Updated the resolver-override to include existing file from the resolver folder

fix #2049