fix(graphql-transformer-core): dont incl in custom stacks stack mapping
GraphQL transformer splits resources into multiple stack to get over the cloudformation limit on the
number of resource a stack can have. If a resource gets moved to different stack it would cause
cloudformation to delete and recreate the same  resource. To prevent this transformer tries to put
the resources in the same stack in subsequent deployment. We used to include custom resources in
the stack mapping which should not be the case as these are custom stacks. Updated the code not to
include custom stack in stack mapping computation.

fix #2167