fix(amplify-codegen): support multi os team workflow in codegen (#2212)
Codegen generates grpaql configuration file which contains paths for schema and documents. These
paths used to be platform specific path which caused issues when the same projects were used in
different Operating Systems. Updated codegen to normalize path to posix style path as it is
supported by all the platforms

fix #2147 and #2002