fix(amplify-category-api): safeguard prompt with empty options
when the CLI prompt the user to select from some options, it uses the inquirer library, and if the
options are empty, inquirer throws error, which leads the CLI to exit without proper messages given
to the user. This PR safeguard against such situation for the RDS selection, and it will print out
proper error message before the CLI exits.

fix #2423