fix: mitigate push failuer when graphql api backend is missing (#2559)
* fix: mitigate push failuer when graphql api backend is missing

Cloud formation push fails in amplify console when a project has GraphQL API but changes are in non
GraphQL API resource. When pushing resources to cloud, Amplify CLI expects the api build directory
to present. When the push is running in the console,  API build directory will be missing because of
the gitignore rule. Updating the build steps to generate build folder when they are missing

https://github.com/aws-amplify/amplify-console/issues/10