fix(graphql-key-transformer): update deleteInput logic to hadle ID
DynamoDB transformer generates all the inputs with the assumption that evey model will be using id
as hash key. But key trasnformer can change this by using a composite key. When a composite key is
used, key transformer updates the mutation inputs to include the composite key in the input. The
mutation input could have fields added by other transformes which we will have to include in the
input, but transformer should delete the field `id:ID!` as this is blindly added by the dynamo db
transformr. Updating the code to handle this special case