fix(amplify-appsync-simulator): support early return from resolvers (#2497)
AppSync has support to return early from the resolver, if the template has #return statement. If the
#return statement is executed in request mapping template, niether the invoker or response template
is executed. If the #return statemnt is executed in an AppSync function, the function returns
immediately with out executing the remaining template or data invoker. Update the mock
impletentation to support this behavior

fix #2427