fix(amplify-provider-awscloudformation): fix template not found
When pushing resources, previously the code would search for cloudformation templates in the build/ folder, if the build folder exists.
This was intended for use by the API resource, but has the unintended effect of causing templates to not be found in other resource folders,
which may contain build folders but still hold their template file in the root resource directory.

To reproduce this issue, create a function and inside the function resource directory create a build/ folder.
When pushing, the cloudformation file will not be found and uploaded to S3.

The resolve the issue, if a cloudformation file is not found inside the build folder, continue to search for templates in the root directory.