fix(amplify-category-function): prevent overwriting of parameters
Function update walkthroguh used to overwrite everything in parameter.json and
function-parameter.json when an update was done. Update flow does not account for all the
configuration options function can have and causes deletion of cognito trigger configuration.
Updated this to retain last configuration and only update the parts that have changed

fix #4065