enhance(amplify-provider-awscloudformation): support multiple Lambdas in a function resource
There may be reasons to have a multiple lambda function resources inside a single function resource. This may be because the function is doing one-off deployment tasks as a custom resource, or handling other processing duties related to the main lambda. For example, one lambda may handle domain logic and handle CRUD for a dynamodb table, another Lambda may do the stream processing of the dynamodb table, and another might setup dynamodb ttl properties, or perform a data migration. At the same time, the details of these functions do not need to be exposed as a top level function and are best served by being encapsulated in another resource. Having the top level resources littered with tiny lambdas can lead to messy code.

This PR changes the push resources code to update all resources in a template that are a lambda and have a Code property. Previously, only the resource named LambdaFunction would have its Code property updated.

Users would not set a Code property in a resource template that they do not expect to have overwritten, but it is technically possible.

There are no better alternatives to handle this use case as the cloudformation provider is in control of the bucket and key names. Users could inline their functions inside the template, but that excludes a lot of nice tooling.