fix(amplify-nodejs-function-runtime-provider): invoke waits close event before rejects
execa module waits untill all output is drained, but it can sends the 'close' event after the the
promise is rejected. Or the promise is rejected after close event but before before functions
finishs. The change will upto 2 seconds to event close is been sent and waits until the close call
back is completed

fix #4439