fix(amplify-provider-awscloudformation): apigw resource download
Extract-zip package was updated to v2 which removed support for
callback based zip extraction. The API GW model download code was
not updated to adhere to new signature. Changed the method to use
the new extract zip signature.
To detect any issues in future updated the the module and dependecies
to Typescript to catch these erros at compile time.

fix #5557