fix(amplify-provider-awscloudformation): apigw resource download (#5564)
Extract-zip package was updated to v2 which removed support for callback based zip extraction. The API GW model download code was not updated to adhere to the new signature. Changed the method to use
the new extract zip signature. To detect any issues in future updated the the module and dependencies to Typescript to catch these errors at compile time.

fix #5557