fix(graphql-auth-transformer): protection for connection types
When generating auth directives to protect fields decorated by @connection we now protect the
operation including its return type instead of protecting only the field. This is important for the
case where queries are not defined on the nested model type since they wont be generated for the
schemas Query.operation.

"fix #4874", "fix #5510"