feat: add support for multiple @key changes in same @model (#6044)
Updated CLI to support multiple changes to @key directive in same @model during update. The CLI will
split the cloudformation stack into mulitple batches of update such that GSIs are added and removed
in separate pushes which out causing any cloudfomration deployment error. This change is behind
FeatureFlage `graphqltransformer.enableIterativeGSIUpdates` which is turned off by default.

The changes supported after enabling this GSI are
1. Rename of @key
2. Add and Remove multiple @keys at the same time in a single model
3. Add and remove fileds to @key

The change only helps in deployment of key changes and does not backfill the data if
the key is a composite key or remove column from Dynamo DB records when a key removed

Co-authored-by: Attila Hajdrik <attila@eyedea.hu>