fix: adding JSON stringify to example python function
We need a JSON string for the body in order to access it in our response message. See this comment: https://github.com/aws-amplify/amplify-js/issues/6390#issuecomment-740757548

I had this issue with my own function which was returning null on my JSON.stringify(response). I found the above comment and it made it work! i.e. I could access the return body.