fix(amplify-provider-awscloudformation): specify 127.0.0.1 as hostname
CLI uses express server in `amplify-provider-awscloudformation` to listen to POST request from
Admin UI login page when `amplify configure`. The default of the hostname is localhost, 0.0.0.0
and 127.0.0.1 on macOS, but Ubuntu on WSL on Windows uses only localhost by default.
Since 127.0.0.1 is used in Admin UI, we should specify 127.0.0.1 as the hostname in
CLI so that CLI can finish `amplify configure`.

Fixes: https://github.com/aws-amplify/amplify-cli/issues/6779
Fixes: https://github.com/aws-amplify/amplify-adminui/issues/134