fix(amplify-category-function): add condition for info printing (#6967)
Description of changes

This PR adds a small condition to the step which prints info when a lambda function resource permissions are updated. This is in reference to my comment here [1].


Description of how you validated changes

Tried removing resource access permissions from a lambda when lambda had access to only 1 resource. This means that after removing resource permissions the lambda should not have access to any resource attributes as environment variables and hence the line should not be printed as info.

[1] - https://github.com/aws-amplify/amplify-cli/pull/6766#issuecomment-805291533