fix: ensure policy resource name when pushing REST APIs (#7192)
* fix: ensure policy resource name when pushing REST APIs

It is possible that older REST APIs have never been updated to
include the policyResourceName field on paths. This commit
ensures that the field is present.

Refs: https://github.com/aws-amplify/amplify-cli/issues/7190

* fix: process CFN and parameters.json separately

When removing the auth and unauth roles from REST APIs, the
CFN template and the parameters.json file should be processed
separately in case they have gotten out of sync. Otherwise, the
generated root stack is likely to have errors.

* fix: ignore AdminQueries during REST API consolidation

This commit moves the check for AdminQueries from
consolidateApiGatewayPolicies() to loadApiWithPrivacyParams(),
as the latter is also called from push-resources.ts.

* fix: add AdminQueries to multiple REST API test

* fix: don't throw on malformed path

Co-authored-by: Colin Ihrig <colihrig@amazon.com>