fix: add sharedId in externalAuthEnable()
This commit updates externalAuthEnable() to include the auth
sharedId when creating an auth resource. The sharedId is given
the lowest priority in the authProps object so that it will not
overwrite any previously existing sharedId. It also uses the
sharedId from cognito-defaults for consistency. This change
prevents the creation of "snsundefined" roles.