fix: prevent infinite loop (#8540)
The loop in getResourcesToBeCreated() pushes items into the
array that it is iterating over. However, there was no check
for cycles, meaning that circular dependencies pushes resources
into the array until an out of memory error occurs. This commit
adds a check to prevent adding the same item to the array more
than once.

Co-authored-by: Ihrig <colihrig@88665a058d28.ant.amazon.com>