ci: breaking change empty commit (#8844)
This PR introduces overrides and export. Overrides enables customers to
extend or modify their resource categories. Export enables them to export their backend
and use it in a CDK app.

BREAKING CHANGE: The breaking change is in the workflow of auth and storage. Auth and Storage
categories can be overriden