fix(amplify-provider-awscloudformation): use the right profile with SSO and credential_process
Make sure the right profile is used when using credential_process

When using credential_process in the AWS profile (to get SSO to work for example) the
ProcessCredentials Provider is used. When using the default chain from the aws-sdk no options are
passed to it and it default to using AWS_PROFILE.

With this change, we detect ```credential_process``` and specifically pass the profile name into the
credential chain to ensure it gets used.

re #4488