fix: update dep and use node test environment (#9434)
There is an out of date dependency on
amplify-provider-awscloudformation, which is pulling in a
number of older dependency versions. This commit fixes the
dependency version. This commit also updates some of our test
suites to use the Node testing environment instead of the
default of jsdom. This is done because the default does not
properly handle some Node globals.

Refs: https://github.com/panva/jose/issues/307

Co-authored-by: Colin Ihrig <colihrig@amazon.com>