Removed /doc/ from transformer
Removed /doc from transformer tool because it conflicts with OS version >= 1.1 and breaks v2 searchable queries. Migration moves you to OpenSearch.

https://docs.amplify.aws/cli/graphql/custom-business-logic/#extend-amplify-generated-resolvers

This part of the docs states that you can extend generated resolvers, if this merge request isn't approved my guess is this could be a valid method to change the indexPath for users who use OS 1.1