fix: remove functionName from transform-host lambdas (#9491)
The @searchable and @predictions directives create a Lambda
function via context.api.host.addLambdaFunction(). This commit
updates addLambdaFunction() to not set the optional functionName
in the CFN. This lets CFN generate a unique name.

The mock functionality also needed to be updated because mock
expects all Lambdas to have this optional property. If the CFN
does not contain a functionName, fall back to the resource name.

Co-authored-by: Colin Ihrig <colihrig@amazon.com>