build: add script to check if the versions are already published (#9734)
Added helper scripts to run a sanity check before starting the release. The script will run learn version and bump the version according to conventional commit and then check if the bumped versions are already published to NPM or if there is a git tag for that version due to a failed publishing. This script needs to be run before starting the release by executing either yarn run check-version-conflicts:beta or check-version-conflicts:release before starting a release