fix: include default values in index arg validation
Create and update mutations generate a VTL snippet that verifies
that partial sort keys are not provided. However, this snippet
only checks for the arguments in $ctx.args.input, which does
not include default or computed values. These values are available
alongside the user inputs in a variable called $mergedValues.
This commit updates the argument checking logic to use the
merged values instead of the raw input values.

Fixes: https://github.com/aws-amplify/amplify-cli/issues/9666