fix(amplify-category-analytics): ensure pinpoint permissions are created
When an AppId is provided the analytics module is not responsible for creating the pinpoint
resource. There is a condition named ```ShouldCreatePinpointApp``` to check for this.

The auth and unAuth roles had this permission specified, which is unnecessary as the resources
within them already have an if which outputs different values based on this condition.