Uploads metadata files required by Studio to the deployment bucket root
Problem:
Amplify Studio is currently using the unzipper library to stream
metatdata files out of the zip file in the customer's deployment bucket
in order to retrieve data it needs for deployments and/or display in the
UI. However, this library is out of active development and has
significant error handling bugs that causing the Amplify Backend control
plane availability to degrade.

Solution:
Upload the needed metadata files to the s3 bucket outside of the zip
file so they are accessible with a single GetObject call and reaching
into the zip file is no longer required.

Testing Done:
yarn setup-dev
amplify-dev init
amplify-dev add auth
amplify-dev push
amplify-dev remove auth
amplify-dev push