cb: enable basic e2e test coverage (#12601)
This adds several items to the CodeBuild setup including CodeCov updates and V8 supported coverage for e2e tests.

The CodeBuild updates include the function _uploadCoverageLinux in the shared-scripts.sh file. This setup utilizes the CodeCov Token secret, managed by SecretsManager and injected by the build. It then gets the latest CodeCov uploader executable and verifies it via SHA. Once the verification is complete, the uploader API is called, automatically searching for coverage artifacts to upload. This update also includes an update to the codecov.yml file to conform to CodeCov validation.

Lastly, in this update we take advantage of the embedded ability of NodeJS to report coverage metrics: https://blog.npmjs.org/post/178487845610/rethinking-javascript-test-coverage.html. We then run the c8 tool to convert that coverage into something that both CodeCov and CodeBuild can consume. It also integrates these reports in CodeBuild directly.